// github-user-search/src/types.ts
export interface User {
  login: string;
  avatar_url: string;
  html_url: string;
  id: number;
  repos_url: string;
  name: string;
  company: string;
  blog: string;
  location: string;
  bio: string;
  twitter_username: string;
  public_repos: number;
  followers: number;
  following: number;
  created_at: string;
}